Ticket #5222 - apply Karim's comments - part 1
