+++ app/models/concerns/annotation_base.rb
@@ -24,8 +24,6 @@ module AnnotationBase
           query = self.annotation_query(type, context)
           klass = (type.blank? || type.is_a?(Array)) ? Annotation : type.camelize.constantize
           relation = klass.where(query)
-          #relation = relation.where(context_id: nil) if context == 'none'
-          #relation = relation.where.not(context_id: nil) if context == 'some'
           relation.order('id DESC')
         end
       end
